feat: restructure component folders
