+++ src/App.tsx
@@ -1,13 +1,13 @@
-import Navbar from "./components/Navbar";
+import Navbar from "./components/Navbar/Navbar";
+import CardsContainer from "./components/CardsContainer/CardsContainer";
+import HeroSection from "./components/Hero/Hero";
 
 function App() {
   return (
     <>
-      <Navbar /> {/* Add the Navbar at the top */}
-      <div className="mt-16">
-        <h1 className="text-7xl text-center">zubi</h1>
-        <p className="text-center mt-4">bridge your skills</p>
-      </div>
+      <Navbar /> 
+      <HeroSection />
+      <CardsContainer/>
     </>
   );
 }
