+++ src/components/Hero/Hero.tsx
@@ -1,0 +1,10 @@
+function HeroSection() {
+  return (
+    <div className="mt-16">
+      <h1 className="text-7xl text-center">Zubi</h1>
+      <p className="text-center mt-4">Bridge your skills</p>
+    </div>
+  );
+}
+
+export default HeroSection;
