+++ src/components/Logo/Logo.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import logo from '../assets/logo.svg';
+import logo from '../../assets/logo.svg';
 
 const Logo: React.FC<{ className?: string }> = ({ className }) => {
   return (
